+++ classes/DataContainer/ModuleContainer.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2019 Heimrich & Hannot GmbH
+ *
+ * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
+ * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ */
+
+
+namespace HeimrichHannot\Banner\DataContainer;
+
+
+class ModuleContainer
+{
+    const DISPLAY_DEFAULT = "default";
+    const DISPLAY_ROTATORY = 'rotatory';
+}