+++ dca/tl_module.php
@@ -9,13 +9,30 @@
  * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
  */
 
-$dc = &$GLOBALS['TL_DCA']['tl_module'];
+$table = 'tl_module';
+$dc = &$GLOBALS['TL_DCA'][$table];
 
-if (in_array('slick', \ModuleLoader::getActive()))
+if (in_array('slick', \Contao\ModuleLoader::getActive()))
 {
 	$dc['palettes']['slick_newslist'] = str_replace(
 		'skipFirst;',
-		'skipFirst;{banner_legend},banner_hideempty,banner_firstview;banner_categories,banner_template;banner_redirect;guests,protected,banner_useragent;',
+		'skipFirst;{banner_legend},banner_hideempty,banner_firstview,banner_categories,banner_template,banner_plus_displayFormat,banner_redirect,guests,protected,banner_useragent;',
 		$dc['palettes']['slick_newslist']
 	);
 }
+
+$dc['fields']['banner_plus_displayFormat'] = [
+    'label'     => &$GLOBALS['TL_LANG'][$table]['banner_plus_displayFormat'],
+    'exclude'   => true,
+    'inputType' => 'select',
+    'options'   => [
+        \HeimrichHannot\Banner\DataContainer\ModuleContainer::DISPLAY_DEFAULT,
+        \HeimrichHannot\Banner\DataContainer\ModuleContainer::DISPLAY_ROTATORY,
+    ],
+    'reference' => &$GLOBALS['TL_LANG'][$table]['banner_plus_displayFormat'],
+    'eval'      => ['includeBlankOption' => true, 'tl_class' => 'w50 clr', 'helpwizard'=>true,],
+    'sql'       => "varchar(10) NOT NULL default ''",
+    'explanation'   => 'banner_plus_displayFormat'
+];
+
+$dc['fields']['banner_redirect']['eval']['tl_class'] .= ' clr';