+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2019-01-09
+
+### Added
+- option to display banners in slick slider rotatary
+
+### Changed
+- increased minimum php version to 5.6
+
 ## [1.0.6] - 2017-04-21
 
 ### Fixed